refactor :: project name change
