+++ app/src/androidTest/java/com/example/mogak_android/ExampleInstrumentedTest.kt
@@ -1,4 +1,4 @@
-package com.example.wsh_aos
+package com.example.mogak_android
 
 import androidx.test.platform.app.InstrumentationRegistry
 import androidx.test.ext.junit.runners.AndroidJUnit4
