+++ app/src/main/AndroidManifest.xml
@@ -10,13 +10,13 @@
         android:label="@string/app_name"
         android:roundIcon="@mipmap/ic_launcher_round"
         android:supportsRtl="true"
-        android:theme="@style/Theme.WSHAOS"
+        android:theme="@style/Theme.Mogak"
         tools:targetApi="31">
         <activity
-            android:name=".MainActivity"
+            android:name="com.example.mogak_android.MainActivity"
             android:exported="true"
             android:label="@string/app_name"
-            android:theme="@style/Theme.WSHAOS">
+            android:theme="@style/Theme.Mogak">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
