+++ app/src/main/java/com/example/mogak_android/MainActivity.kt
@@ -1,4 +1,4 @@
-package com.example.wsh_aos
+package com.example.mogak_android
 
 import android.os.Bundle
 import androidx.activity.ComponentActivity
@@ -10,13 +10,13 @@ import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.tooling.preview.Preview
-import com.example.wsh_aos.theme.WSHAOSTheme
+import com.example.mogak_android.theme.WSHAOSTheme
 
 class MainActivity : ComponentActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         setContent {
-            com.example.wsh_aos.theme.WSHAOSTheme {
+            WSHAOSTheme {
                 // A surface container using the 'background' color from the theme
                 Surface(
                     modifier = Modifier.fillMaxSize(),
@@ -40,7 +40,7 @@ fun Greeting(name: String, modifier: Modifier = Modifier) {
 @Preview(showBackground = true)
 @Composable
 fun GreetingPreview() {
-    com.example.wsh_aos.theme.WSHAOSTheme {
+    WSHAOSTheme {
         Greeting("Android")
     }
 }
