+++ app/src/main/java/com/example/mogak_android/MogakApp.kt
@@ -1,4 +1,4 @@
-package com.example.wsh_aos
+package com.example.mogak_android
 
 class MogakApp {
 }
