+++ app/src/main/java/com/example/mogak_android/MogakApplication.kt
@@ -1,4 +1,4 @@
-package com.example.wsh_aos
+package com.example.mogak_android
 
 import android.app.Application
 import com.example.di.apiModule
