+++ app/src/main/java/com/example/mogak_android/navigation/AuthNavigation.kt
@@ -1,0 +1,4 @@
+package com.example.mogak_android.navigation
+
+class AuthNavigation {
+}
