+++ app/src/main/java/com/example/mogak_android/theme/Color.kt
@@ -1,4 +1,4 @@
-package com.example.wsh_aos.theme
+package com.example.mogak_android.theme
 
 import androidx.compose.ui.graphics.Color
