+++ app/src/main/res/values/strings.xml
@@ -1,3 +1,3 @@
 <resources>
-    <string name="app_name">WSH-AOS</string>
+    <string name="app_name">Mogak</string>
 </resources>
